implement protection of create, update and, delete operations

+++ routes/house.routes.js
@@ -5,23 +5,21 @@ const House = require("../models/House.model");
 const Reservation = require("../models/Reservation.model");
 
 const { isAuthenticated } = require("../middleware/jwt.middleware.js");
-
+const attachCurrentUser = require("../middleware/attachCurrentUser");
 // POST /api/houses  -  Creates a new house
-router.post("/houses",isAuthenticated, (req, res, next) => {
+router.post("/houses", isAuthenticated, attachCurrentUser, (req, res, next) => {
   const { title, description, cost, location } = req.body;
-
   const newHouse = {
     title,
     description,
     cost,
     location,
+    ownerId: req.currentUser._id,
     reservations: [],
   };
 
   if (!title) {
-    return res
-       .status(400)
-       .json({ message: "Please provide the title" });
+    return res.status(400).json({ message: "Please provide the title" });
   }
 
   House.create(newHouse)
@@ -71,30 +69,53 @@ router.get("/houses/:houseId", (req, res, next) => {
 });
 
 // PUT  /api/houses/:houseId  -  Updates a specific house by id
-router.put("/houses/:houseId", isAuthenticated, (req, res, next) => {
-  const { houseId } = req.params;
+router.put(
+  "/houses/:houseId",
+  isAuthenticated,
+  attachCurrentUser,
+  (req, res, next) => {
+    const { houseId } = req.params;
 
-  if (!mongoose.Types.ObjectId.isValid(houseId)) {
-    res.status(400).json({ message: "Specified id is not valid" });
-    return;
-  }
+    if (!mongoose.Types.ObjectId.isValid(houseId)) {
+      res.status(400).json({ message: "Specified id is not valid" });
+      return;
+    }
 
-  House.findByIdAndUpdate(houseId, req.body, { new: true })
-    .then((updatedHouse) => {
-      console.log(updatedHouse);
-      res.json(updatedHouse);
-    })
-    .catch((err) => {
-      console.log("error updating a house...", err);
-      res.status(500).json({
-        message: "error updating a house",
-        error: err,
+    House.findById(req.params.houseId)
+      .then((houseDetails) => {
+        if (
+          req.currentUser._id &&
+          req.currentUser._id.equals(houseDetails.ownerId)
+        ) {
+          House.findByIdAndUpdate(houseId, req.body, { new: true })
+            .then((updatedHouse) => {
+              console.log(updatedHouse);
+              res.json(updatedHouse);
+            })
+            .catch((err) => {
+              console.log("error updating a house...", err);
+              res.status(500).json({
+                message: "error updating a house",
+                error: err,
+              });
+            });
+        } else {
+          res.status(400).json({
+            message:
+              "Users can updated only thier own data.Please Check the data you are updating!",
+          });
+          return;
+        }
+      })
+      .catch((err) => {
+        console.log("error getting house details from DB", err);
+        next();
       });
-    });
-});
+  }
+);
 
 // DELETE  /api/houses/:houseId  -  Deletes a specific house by id
-router.delete("/houses/:houseId",isAuthenticated, (req, res, next) => {
+router.delete("/houses/:houseId", isAuthenticated,attachCurrentUser, (req, res, next) => {
   const { houseId } = req.params;
 
   if (!mongoose.Types.ObjectId.isValid(houseId)) {
@@ -102,16 +123,36 @@ router.delete("/houses/:houseId",isAuthenticated, (req, res, next) => {
     return;
   }
 
-  House.findByIdAndRemove(houseId)
-    .then(() =>
-      res.json({ message: `House with ${houseId} is removed successfully.` })
-    )
+  House.findById(req.params.houseId)
+    .then((houseDetails) => {
+      if (
+        req.currentUser._id &&
+        req.currentUser._id.equals(houseDetails.ownerId)
+      ) {
+        House.findByIdAndRemove(houseId)
+          .then(() =>
+            res.json({
+              message: `House with ${houseId} is removed successfully.`,
+            })
+          )
+          .catch((err) => {
+            console.log("error deleting a house...", err);
+            res.status(500).json({
+              message: "error deleting a house",
+              error: err,
+            });
+          });
+      } else {
+        res.status(400).json({
+          message:
+            "Users can delete only thier own data.Please Check the data you are deleting!",
+        });
+        return;
+      }
+    })
     .catch((err) => {
-      console.log("error deleting a house...", err);
-      res.status(500).json({
-        message: "error deleting a house",
-        error: err,
-      });
+      console.log("error deleting house details from DB", err);
+      next();
     });
 });
